refactor: extract LandingScenarioSections to resolve re-rendering issues
